Move useNetworkConnectivity hook from @wordpress/compose to @wordpress/editor package
