+++ packages/editor/src/components/offline-status/index.native.js
@@ -6,11 +6,7 @@ import { AccessibilityInfo, Text, View } from 'react-native';
 /**
  * WordPress dependencies
  */
-import {
-	usePreferredColorSchemeStyle,
-	useNetworkConnectivity,
-	usePrevious,
-} from '@wordpress/compose';
+import { usePreferredColorSchemeStyle, usePrevious } from '@wordpress/compose';
 import { Icon } from '@wordpress/components';
 import { offline as offlineIcon } from '@wordpress/icons';
 import { __ } from '@wordpress/i18n';
@@ -20,6 +16,7 @@ import { useEffect } from '@wordpress/element';
  * Internal dependencies
  */
 import styles from './style.native.scss';
+import useNetworkConnectivity from '../../hooks/use-network-connectivity';
 
 /**
  * Conditionally announces messages for screen reader users. This Hook provides
