+++ packages/editor/src/hooks/index.native.js
@@ -1,0 +1,4 @@
+/**
+ * Internal dependencies
+ */
+import './use-network-connectivity';
