+++ packages/editor/src/index.native.js
@@ -12,5 +12,6 @@ import './hooks';
 
 export { store } from './store';
 export * from './components';
+export * from './hooks';
 export * from './utils';
 export * from './private-apis';
